refactor(shorebird_cli): speedup aar differ by using same zip diff technique as ipa
