+++ packages/shorebird_cli/lib/src/archive_analysis/android_archive_differ.dart
@@ -1,26 +1,8 @@
-import 'package:collection/collection.dart';
 import 'package:path/path.dart' as p;
 import 'package:shorebird_cli/src/archive_analysis/archive_analysis.dart';
 import 'package:shorebird_cli/src/archive_analysis/archive_differ.dart';
 
 abstract class AndroidArchiveDiffer extends ArchiveDiffer {
-  @override
-  bool containsPotentiallyBreakingAssetDiffs(FileSetDiff fileSetDiff) {
-    final assetsDiff = assetsFileSetDiff(fileSetDiff);
-
-    // If assets were added, we need to warn the user about asset differences.
-    if (assetsDiff.addedPaths.isNotEmpty) {
-      return true;
-    }
-
-    return assetsDiff.changedPaths
-        .whereNot(
-          (path) =>
-              ArchiveDiffer.assetFileNamesToIgnore.contains(p.basename(path)),
-        )
-        .isNotEmpty;
-  }
-
   @override
   bool containsPotentiallyBreakingNativeDiffs(FileSetDiff fileSetDiff) =>
       nativeFileSetDiff(fileSetDiff).isNotEmpty;
